// src/types/CollapseState.ts
export enum CollapseState {
    NONE = "NONE",
    TIMEFRAME_SELECT = "TIMEFRAME_SELECT",
    DATE_TIME_SELECT = "DATE_TIME_SELECT",
    OUTBOUND_TOKEN_SELECT = "OUTBOUND_TOKEN_SELECT",
    INBOUND_TOKEN_SELECT = "INBOUND_TOKEN_SELECT",
    WRAP_TOKENS = "WRAP_TOKENS",
    SWAP_APPROVE = "SWAP_APPROVE",
    SWAP_SUBMITTING = "SWAP_SUBMITTING",
    SWAP_SUCCESS = "SWAP_SUCCESS",
    SWAP_FAILURE = "SWAP_FAILURE",
}